+++ prova.py
@@ -1,4 +1,4 @@
-from data_retriever import data_retriever
+from service import data_retriever
 
 
 print(data_retriever.retrieve_country_from_csv('Italy'))
